refactor: moved classes to separate files
