+++ AdSecGH/Helpers/BusinessExtensions.cs
@@ -10,14 +10,11 @@ using Grasshopper.Kernel.Types;
 
 using Oasys.Business;
 
-using OasysGH;
-using OasysGH.Components;
-
 using Attribute = Oasys.Business.Attribute;
 
 namespace Oasys.GH.Helpers {
 
-  public static class Populator {
+  public static class BusinessExtensions {
 
     private static readonly Dictionary<Type, Func<Attribute, IGH_Param>> ToGhParam
       = new Dictionary<Type, Func<Attribute, IGH_Param>> {
@@ -91,75 +88,6 @@ namespace Oasys.GH.Helpers {
     }
   }
 
-  public abstract class BusinessOasysDropdownGlue<T> : GH_OasysDropDownComponent, IDefaultValues
-    where T : IBusinessComponent {
-    private readonly T _businessComponent = Activator.CreateInstance<T>();
-
-    public BusinessOasysDropdownGlue() : base("", "", "", "", "") {
-      Name = _businessComponent.Metadata.Name;
-      NickName = _businessComponent.Metadata.NickName;
-      Description = _businessComponent.Metadata.Description;
-      Category = _businessComponent.Organisation.Category;
-      SubCategory = _businessComponent.Organisation.SubCategory;
-    }
-
-    public override Guid ComponentGuid { get; }
-
-    public override OasysPluginInfo PluginInfo { get; }
-
-    public void SetDefaultValues() { _businessComponent.SetDefaultValues(this); }
-
-    protected override void RegisterInputParams(GH_InputParamManager pManager) {
-      _businessComponent.PopulateInputParams(this);
-    }
-
-    protected override void RegisterOutputParams(GH_OutputParamManager pManager) {
-      _businessComponent.PopulateOutputParams(this);
-    }
-
-    public override void SetSelected(int i, int j) { }
-
-    protected override void SolveInternal(IGH_DataAccess da) {
-      _businessComponent.Compute();
-      _businessComponent.SetOutputValues(this, da);
-    }
-
-    protected override void InitialiseDropdowns() {
-      _spacerDescriptions = new List<string>();
-      _dropDownItems = new List<List<string>>();
-      _selectedItems = new List<string>();
-      _isInitialised = true;
-    }
-  }
-
-  public abstract class BusinessOasysGlue<T> : GH_OasysComponent, IDefaultValues where T : IBusinessComponent {
-
-    private readonly T _businessComponent = Activator.CreateInstance<T>();
-
-    public BusinessOasysGlue(string name, string nickname, string description, string category, string subCategory) :
-      base(name, nickname, description, category, subCategory) { }
-
-    public override OasysPluginInfo PluginInfo { get; } = AdSecGH.PluginInfo.Instance;
-    public void SetDefaultValues() { _businessComponent.SetDefaultValues(this); }
-
-    protected override void RegisterInputParams(GH_InputParamManager pManager) {
-      _businessComponent.PopulateInputParams(this);
-    }
-
-    protected override void RegisterOutputParams(GH_OutputParamManager pManager) {
-      _businessComponent.PopulateOutputParams(this);
-    }
-
-    protected override void SolveInstance(IGH_DataAccess DA) {
-      _businessComponent.Compute();
-      _businessComponent.SetOutputValues(this, DA);
-    }
-  }
-
-  public interface IDefaultValues {
-    void SetDefaultValues();
-  }
-
   public class DummyBusiness : IBusinessComponent {
 
     public DoubleParameter Alpha { get; set; } = new DoubleParameter {
@@ -221,7 +149,7 @@ namespace Oasys.GH.Helpers {
     protected override void SolveInstance(IGH_DataAccess DA) { }
   }
 
-  public class DummyOasysDropdown : BusinessOasysDropdownGlue<DummyBusiness> {
+  public class Dummy : BusinessGlue<DummyBusiness> {
     public override GH_Exposure Exposure { get; } = GH_Exposure.hidden;
     public override Guid ComponentGuid => new Guid("CAA08C9E-417C-42AE-B704-91F214C8C871");
   }
